Removed permission checks from Image shortcode
Should be handled by the file routing, rather than at render time.
We have to assume that the output of HTMLText (e.g. SiteTree.Content)
is cached, so we can't have user/session specific behaviour on it.

If the image can't be found (at render time), the image tag is removed from the output,
rather than rendering it with an outdated 'src' attribute.